refactor: 수정, 삭제에 맞는 drawerFooter 그리기, modifySchedule api 연동
