+++ src/containers/home/ScheduleDrawer/pages/CreateSchedulePage/index.tsx
@@ -1,0 +1,7 @@
+import { Alert, Snackbar, Stack } from "@mui/material";
+
+function CreateSchedulePage() {
+  return <></>;
+}
+
+export default CreateSchedulePage;
